vk: move flipping buffer impl to std/flipping.c
